+++ packages/libs/datastore/src/file-system/index.ts
@@ -1,0 +1,4 @@
+export * from "./file-system";
+export * from "./memory-file-system";
+export * from "./real-filesystem";
+export * from "./enhanced-filesystem";
